+++ src/controllers/favorite.controller.ts
