+++ src/controllers/index.ts
@@ -1,4 +1,4 @@
-import { createFavorites } from './fav.controller'
+import { createFavorites } from './favorite.controller'
 import { login, register } from './auth.controller'
 
 export { login, register, createFavorites }
