+++ src/routes/favorite.route.ts
@@ -1,7 +1,7 @@
 import { createFavorites } from '../controllers'
 import { FavoriteSchema } from '../schemas'
 import { Router } from 'express'
-import { validateResult, isAuthenticated } from './../middlewares'
+import { validateResult, isAuthenticated } from '../middlewares'
 
 export const favRouter = Router()
